refactor: sale-detail 추천·관련 팝업 에러 처리 및 훅 경로 통일
